Added more security, now, only authorized admins can view and edit patient data

+++ pages/api/patients/[id].js
@@ -9,13 +9,14 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
-    cookies,
   } = req;
 
+  let token = req.headers.cookie
+
   if (method === 'GET') {
-    // if (!token || token !== process.env.TOKEN) {
-    //   return res.status(401).json('Not authenticated');
-    // }
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('Not authenticated');
+    }
 
     try {
       const patient = await Patient.findById(id);
@@ -26,9 +27,9 @@ export default async function handler(req, res) {
   }
 
   if (method === 'DELETE') {
-    // if (!token || token !== process.env.TOKEN) {
-    //   return res.status(401).json('Not authenticated');
-    // }
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('Not authenticated');
+    }
 
     try {
       await Patient.findByIdAndDelete(id);
